LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/translate-composition-expected.txt


FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [-50px 50%] to add [100%] at (-1) should be [-100% calc(200px + 100%) 300px] assert_equals: expected "0 200 300 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [-50px 50%] to add [100%] at (0) should be [calc(50px + 0%) calc(200px + 50%) 300px] assert_equals: expected "50 200 300 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [-50px 50%] to add [100%] at (0.25) should be [calc(62.5px + 25%) calc(200px + 37.5%) 300px] assert_equals: expected "62.5 200 300 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [-50px 50%] to add [100%] at (0.75) should be [calc(87.5px + 75%) calc(200px + 12.5%) 300px] assert_equals: expected "87.5 200 300 " but got "100 200 300 "
PASS Compositing: property <translate> underlying [100px 200px 300px] from add [-50px 50%] to add [100%] at (1) should be [calc(100px + 100%) calc(200px + 0%) 300px]
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [-50px 50%] to add [100%] at (2) should be [calc(150px + 200%) calc(200px - 50%) 300px] assert_equals: expected "150 200 300 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [50% 100px] to replace [200px 50% 100px] at (-1) should be [100% calc(600px - 50%) 500px] assert_equals: expected "0 600 500 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [50% 100px] to replace [200px 50% 100px] at (0) should be [calc(100px + 50%) calc(300px + 0%) 300px] assert_equals: expected "100 300 300 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [50% 100px] to replace [200px 50% 100px] at (0.25) should be [calc(125px + 37.5%) calc(225px + 12.5%) 250px] assert_equals: expected "125 225 250 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [50% 100px] to replace [200px 50% 100px] at (0.75) should be [calc(175px + 12.5%) calc(75px + 37.5%) 150px] assert_equals: expected "175 75 150 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [50% 100px] to replace [200px 50% 100px] at (1) should be [calc(200px + 0%) 50% 100px] assert_equals: expected "200 0 100 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from add [50% 100px] to replace [200px 50% 100px] at (2) should be [calc(300px - 50%) calc(-300px + 100%) -100px] assert_equals: expected "300 - 300 - 100 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from replace [50% 100px] to add [200px 50% 100px] at (-1) should be [calc(-300px + 100%) -50% -400px] assert_equals: expected "- 300 0 - 400 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from replace [50% 100px] to add [200px 50% 100px] at (0) should be [50%  calc(100px + 0%)] assert_equals: expected "0 100 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from replace [50% 100px] to add [200px 50% 100px] at (0.25) should be [calc(75px + 37.5%) calc(125px + 12.5%) 100px] assert_equals: expected "75 125 100 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from replace [50% 100px] to add [200px 50% 100px] at (0.75) should be [calc(225px + 12.5%) calc(175px + 37.5%) 300px] assert_equals: expected "225 175 300 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from replace [50% 100px] to add [200px 50% 100px] at (1) should be [calc(300px + 0%) calc(200px + 50%) 400px] assert_equals: expected "300 200 400 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [100px 200px 300px] from replace [50% 100px] to add [200px 50% 100px] at (2) should be [calc(600px - 50%) calc(300px + 100%) 800px] assert_equals: expected "600 300 800 " but got "100 200 300 "
FAIL Compositing: property <translate> underlying [none] from replace [none] to add [100px] at (-1) should be [-100px] assert_equals: expected "- 100 0 " but got "none "
PASS Compositing: property <translate> underlying [none] from replace [none] to add [100px] at (0) should be [none]
FAIL Compositing: property <translate> underlying [none] from replace [none] to add [100px] at (0.25) should be [25px] assert_equals: expected "25 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [none] to add [100px] at (0.75) should be [75px] assert_equals: expected "75 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [none] to add [100px] at (1) should be [100px] assert_equals: expected "100 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [none] to add [100px] at (2) should be [200px] assert_equals: expected "200 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from add [none] to add [100px] at (-1) should be [-100px] assert_equals: expected "- 100 0 " but got "none "
PASS Compositing: property <translate> underlying [none] from add [none] to add [100px] at (0) should be [none]
FAIL Compositing: property <translate> underlying [none] from add [none] to add [100px] at (0.25) should be [25px] assert_equals: expected "25 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from add [none] to add [100px] at (0.75) should be [75px] assert_equals: expected "75 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from add [none] to add [100px] at (1) should be [100px] assert_equals: expected "100 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from add [none] to add [100px] at (2) should be [200px] assert_equals: expected "200 0 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to replace [none] at (-1) should be [0px 80px 120px] assert_equals: expected "0 80 120 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to replace [none] at (0) should be [0px 40px 60px] assert_equals: expected "0 40 60 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to replace [none] at (0.25) should be [0px 30px 45px] assert_equals: expected "0 30 45 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to replace [none] at (0.75) should be [0px 10px 15px] assert_equals: expected "0 10 15 " but got "none "
PASS Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to replace [none] at (1) should be [none]
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to replace [none] at (2) should be [0px -40px -60px] assert_equals: expected "0 - 40 - 60 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to add [none] at (-1) should be [0px 80px 120px] assert_equals: expected "0 80 120 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to add [none] at (0) should be [0px 40px 60px] assert_equals: expected "0 40 60 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to add [none] at (0.25) should be [0px 30px 45px] assert_equals: expected "0 30 45 " but got "none "
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to add [none] at (0.75) should be [0px 10px 15px] assert_equals: expected "0 10 15 " but got "none "
PASS Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to add [none] at (1) should be [none]
FAIL Compositing: property <translate> underlying [none] from replace [0px 40px 60px] to add [none] at (2) should be [0px -40px -60px] assert_equals: expected "0 - 40 - 60 " but got "none "
FAIL Compositing: property <translate> underlying [80px 20px] from add [none] to replace [0px 40px 60px] at (-1) should be [160px 0px -60px] assert_equals: expected "160 0 - 60 " but got "80 20 "
PASS Compositing: property <translate> underlying [80px 20px] from add [none] to replace [0px 40px 60px] at (0) should be [80px 20px]
FAIL Compositing: property <translate> underlying [80px 20px] from add [none] to replace [0px 40px 60px] at (0.25) should be [60px 25px 15px] assert_equals: expected "60 25 15 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [none] to replace [0px 40px 60px] at (0.5) should be [40px 30px 30px] assert_equals: expected "40 30 30 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [none] to replace [0px 40px 60px] at (0.75) should be [20px 35px 45px] assert_equals: expected "20 35 45 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [none] to replace [0px 40px 60px] at (1) should be [0px 40px 60px] assert_equals: expected "0 40 60 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [none] to replace [0px 40px 60px] at (2) should be [-80px 60px 120px] assert_equals: expected "- 80 60 120 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [0px 40px 60px] to replace [none] at (-1) should be [160px 120px 120px] assert_equals: expected "160 120 120 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [0px 40px 60px] to replace [none] at (0) should be [80px 60px 60px] assert_equals: expected "80 60 60 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [0px 40px 60px] to replace [none] at (0.25) should be [60px 45px 45px] assert_equals: expected "60 45 45 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [0px 40px 60px] to replace [none] at (0.5) should be [40px 30px 30px] assert_equals: expected "40 30 30 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [0px 40px 60px] to replace [none] at (0.75) should be [20px 15px 15px] assert_equals: expected "20 15 15 " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [0px 40px 60px] to replace [none] at (1) should be [none] assert_equals: expected "none " but got "80 20 "
FAIL Compositing: property <translate> underlying [80px 20px] from add [0px 40px 60px] to replace [none] at (2) should be [-80px -60px -60px] assert_equals: expected "- 80 - 60 - 60 " but got "80 20 "
